Break review summary card into separate components
